# WIP: Implement encode_native() without going through the character vector