--- a/R/native.R
+++ b/R/native.R
@@ -12,10 +12,9 @@
 #' 
 #' @param colour For `encode_native` either a vector of hex-encoded 
 #' colours/colour names or a matrix encoding colours in any of the supported 
-#' colour spaces. If the  latter, the colours will be encoded to a hex string 
-#' using [encode_colour()] first. For `decode_native` it is a vector of 
+#' colour spaces. For `decode_native` it is a vector of 
 #' integers.
-#' @param ... Arguments passed on to [encode_colour()]
+#' @inheritParams encode_colour
 #' 
 #' @return `encode_native()` returns an integer vector and `decode_native()`
 #' returns a character vector, both matching the length of the input.
@@ -33,21 +32,25 @@
 #' # Convert back
 #' decode_native(native_col)
 #' 
-encode_native <- function(colour, ...) {
-  if (!is.character(colour)) {
-    colour <- encode_colour(colour, ...)
+encode_native <- function(colour, alpha = NULL, from = 'rgb', white = 'D65') {
+  if (from != 'rgb') {
+    white <- as_white_ref(white)
   }
-  encode_native_c(colour)
+  encode_c(colour, alpha, colourspace_match(from), white, out_format = colour_format_match("native"))
 }
+
 #' @rdname native_encoding
 #' @export
 decode_native <- function(colour) {
   decode_native_c(colour)
 } 
 
+# This is left for benchmarking the two implementations:
 encode_native_c <- function(colour) {
   .Call(`farver_encode_native_c`, colour)
 }
+
+
 decode_native_c <- function(colour) {
   .Call(`farver_decode_native_c`, as.integer(colour))
 }