… NAMESPACE
… NEWS.md
… man/replace_alpha_native.Rd
… src/encode.h
… src/init.cpp
… tests/testthat/test-comparison.R
… tests/testthat/test-conversion.R
… tests/testthat/test-encoding.R